+++ source/Modules/SqlServerDsc.Common/SqlServerDsc.Common.psm1
@@ -245,8 +245,12 @@ function Invoke-InstallationMediaCopy
         Create a destination folder so the media files aren't written
         to the root of the Temp folder.
     #>
-    $mediaDestinationFolder = Split-Path -Path $SourcePath -Leaf
-    if (-not $mediaDestinationFolder )
+    $serverName, $shareName, $leafs = ($SourcePath -replace '\\\\') -split '\\'
+    if ($leafs)
+    {
+        $mediaDestinationFolder = $leafs -join '\'
+    }
+    else
     {
         $mediaDestinationFolder = New-Guid | Select-Object -ExpandProperty Guid
     }